avoid compiling out test case asserts when not debugging
Adding -DNDEBUG will create lots of unused variable warnings from the
test cases. The ideal solution is to use an alternative assert macro for
the test system, but for now, at least don't actively cause problems.

This assumes -DNDEBUG is not already set in the environment, or you get
to pick up the pieces (unless you're explicitly debugging, and then it
unsets the macro).